requirements.txt
Django==4.0.3
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
requests==2.27.1
six==1.16.0
python-dotenv